requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Markdown==3.3.7
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
